Moved disruption classes to a single package (#3054)
