+++ ambry-prioritization/src/main/java/com/github/ambry/replica/prioritization/disruption/Operation.java
@@ -12,7 +12,7 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  */
 
-package com.github.ambry.replica.prioritization;
+package com.github.ambry.replica.prioritization.disruption;
 
 
 /**
